cub3d.c ile birleştirildi
