+++ Makefile
@@ -26,7 +26,7 @@ fclean : clean
 
 re : fclean all
 
-.PHONY : all clean fclean re
+.PHONY : all clean fclean re deneme testdeneme
 
 norm :
 	norminette
